// src/app/models/user.ts
export class User {
  id: number;
  name: string;
  gender: string;
  address: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  state: string;
  city: string;
  zip: string;
  height: string;
  weight: string;
  presentIssues: string;
  pastIssues: string;
  familyHistoryIssues: string;
  foodChoices: string;
  allergy: string;
  medication: string;
  sleepTime: string;
  foodTime: string;
  dayToDayActivities: string;
  createdBy: string;
  modifiedBy: string;
  createdDate: string;
  modifiedDate: string;
  isActive: boolean;
}